+++ frontend/src/containers/EmployeeFile/auditor/index.jsx
@@ -1,0 +1,5 @@
+const AuditorEmployeeFile = () => {
+  return <div>Auditor</div>;
+};
+
+export default AuditorEmployeeFile;
